pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "release-docs-agent"
version = "0.1.0"
description = "AI agent for automating documentation updates from Jira, Confluence, and Bitbucket"
authors = [{name = "Release Docs Agent", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.16",
    "langchain>=0.3.7",
    "langchain-openai>=0.2.8",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "rich>=13.9.4",
    "atlassian-python-api>=3.41.0",
    "gitpython>=3.1.43",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
]

[project.scripts]
release-docs = "src.app.main:cli"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
